+++ KhoaLuan_QLNhaTro/Models/DetailBill.cs
@@ -9,10 +9,10 @@ namespace KhoaLuan_QLNhaTro.Models
         public string BillId { get; set; }
         public Guid ServiceId { get; set; }
         public int Number {  get; set; }
-        public float Price { get; set; }
-        public float Total {  get; set; }
-        public float? OldNumber { get; set; }
-        public float? NewNumber { get; set; }
+        public decimal Price { get; set; }
+        public decimal Total {  get; set; }
+        public decimal? OldNumber { get; set; }
+        public decimal? NewNumber { get; set; }
         public virtual Bill Bill { get; set; }
         public virtual Service Service { get; set; }
     }
