+++ KhoaLuan_QLNhaTro/Models/Room.cs
@@ -8,10 +8,10 @@ namespace KhoaLuan_QLNhaTro.Models
         [Key]
         public Guid Id { get; set; }
         public string Name { get; set; }
-        public float Price { get; set; }
+        public decimal Price { get; set; }
         public string Status { get; set; }
         public int FLoorNumber { get; set; }
-        public float Acreage { get; set; }
+        public decimal Acreage { get; set; }
         public DateTime CreateAt { get; set; }
         public DateTime UpdateAt { get; set; }
         public Guid? HouseId { get; set; }
