+++ KhoaLuan_QLNhaTro/Models/ViewModel/BillViewModal.cs
@@ -6,10 +6,10 @@ namespace KhoaLuan_QLNhaTro.Models
     {
         public string BillId { get; set; }
         public string RoomName { get; set; }
-        public float RoomPrice { get; set; }
+        public decimal RoomPrice { get; set; }
         public string Status { get; set; }
         public List<DetailBillViewModel> Services { get; set; }
         public List<Room> Rooms { get; set; }
-        public float TotalBill { get; set; }
+        public decimal TotalBill { get; set; }
     }
 }
