+++ KhoaLuan_QLNhaTro/Models/ViewModel/DetailBillViewModel.cs
@@ -4,7 +4,7 @@
     {
         public string ServiceName { get; set; }
         public int Quantity { get; set; }
-        public float UnitPrice { get; set; }
-        public float TotalPrice { get; set; }
+        public decimal UnitPrice { get; set; }
+        public decimal TotalPrice { get; set; }
     }
 }
